edgeless_benchmark: refactor code
