+++ edgeless_benchmark/src/lib.rs
@@ -1,4 +1,6 @@
 // SPDX-FileCopyrightText: © 2023 Claudio Cicconetti <c.cicconetti@iit.cnr.it>
 // SPDX-License-Identifier: MIT
 
+pub mod engine;
 pub mod redis_dumper;
+pub mod workflow_type;
